+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next (YYYY-MM-DD)
+## 0.2.0 (2018-11-15)
 
 - Adds a simple `Matrix` class.
 - Adds a simple `Vector` class.